[VSIC-2] Initial checkin of VSI version 2.0.0 - Dynamic memory managed shared memory.
This is a major update to the VSI code.  The shared memory segment is no
longer a static data structure generated at compile time but a dynamically
allocated storage area managed via calls to "sm_malloc" and "sm_free".  This
version also incorporates a B-tree implementation that uses the shared memory
segment to store data and can be used to model "maps" or "dictionaries"
(depending on your choice of programming language).