Fixed issue with traversing sub directories when variables are passed on the command line
When traversing subdirectories for building the -C <subdir> option for make was used.
For some reason this caused issues when passing a parameter to the command line,
such as FOO=BAR. In that case the toplevel Makefile with its SUBDIRS declaration
was called instead of the one in the sub directory. Replacing the -C <subdir>
with cd <subdir> resolves the issue.

Signed-off-by: Rudolf J Streif <rstreif@jaguarlandrover.com>