Avoiding make built-in variables
The built-in variables of make cause problems if you want to
be able to override them in the makefile as well as with
command line settings. If you use LD = gcc then you cannot
override from the command line. If you use LD ?= gcc then
the default built-in setting of LD, which is ld, prevails.
Passing --no-builtin-variables to make resolves this.

Signed-off-by: Rudolf J Streif <rstreif@jaguarlandrover.com>