Replace all TAB characters with 4 spaces
The source code had inconsistent indentation style, mixing spaces and
tabs.  Spaces were dominant and it is also what I prefer.  It was
possible to derive that a tab length (a.k.a. tabstop) of 4 has been used
so I have therefore assumed that it is the maintainer's desired
indentation step.

In the cases a tab was used in the middle of a line (or anywhere that
text was aligned to a column not divisible by 4), I tried to insert the
correct number of spaces to keep the same look as an editor with
tabstop=4 would show.  For example:

//  Text starts here
  ^^ Here a single tab ends up being shown as 2 spaces if tabstop=4

The code examples in api/README.md were also adjusted in the same way,
so that they are displayed properly regardless of tab size.