Fixed a crash when adding data to a B-tree index.
This was an edge case in the B-tree logic that occurred when a tree node
needed to be split into 2 blocks and the data record at the split point that
gets moved to the parent node was being moved to the last record location in
the root node.  This resulted in the memory location immediately following the
end of the root node being overwritten with the split point record data.  That
eventually caused more memory corruption and the crash.