utils.c: Add guard on LOGs for pthread internals using VSI_DEBUG > 3
After discussion we concluded that logs were likely failing because
they referenced internal struct members that have seemingly changed name
in updated versions of a used library (libpthread, I suppose).

VSI_DEBUG now needs to be set to a value greater than 3 to compile &
execute that part of the logs.  This simple solution was chosen because
I was not sure of the long term intention for these logs.

Consequently, note that setting VSI_DEBUG > 3 compilation is still
likely to fail compilation on many systems, and the code would then need
another update if you're interested in producing those logs.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>