utils.c:  Add guard to LOG statements referring to internal vars
Compilation errors due to the access of non-existing internal variables
for semaphore/mutex, etc.

It is understood that those internal vars are not a public/stable api
and subject to change.  The previous fix in 68bc8a46 compiles it
conditionally depending on the glibc version.  This seemed to fix
x86/x86_64 builds but seemed to fail similarly on the arm platforms
supported by GDP.  (I have not checked whether this is another
architetural or it is using another glibc version).

VSI_DEBUG now needs to be set to a value greater than 3 to compile &
execute that part of the logs. This simple solution was chosen because I
was not sure of the long-term intention for these logs.

Consequently, note that setting VSI_DEBUG > 3 compilation is still
likely to fail compilation on some systems and may need another tweak
first.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>